refactor: 패키지 구조 변경 및 시간 페이지 추가
