+++ README.md
@@ -4,6 +4,8 @@
 
 ### 예약 추가 (POST /reservations)
 
+예약에 필요한 데이터 입력 시 추가 가능
+
 #### 요청 예시
 
 ```json
@@ -17,11 +19,13 @@
 #### 응답 예시
 
 ```json
+Location: "/reservations/1"
+
 {
-  "id": 1,
-  "name": "브라운",
-  "date": "2025-06-30",
-  "time": "15:40"
+"id": 1,
+"name": "브라운",
+"date": "2025-06-30",
+"time": "15:40"
 }
 ```
 
@@ -33,7 +37,7 @@
 | 주어진 날짜 형식에 맞지 않을 경우          | 400 Bad Request | `날짜(년도-월-일)형식에 맞게 입력해 주세요. ex) 2020-12-31` |
 | 주어진 시간 형식에 맞지 않을 경우          | 400 Bad Request | `시간(시:분)형식에 맞게 입력해 주세요. ex) 15:30`         |
 | 과거의 날짜와 시간에 예약 할 경우          | 400 Bad Request | `이미 지난 날짜 및 시간은 예약할 수 없어요.`                |
-| 삭제하려는 예약이 존재하지 않을 경우         | 400 Bad Request | `예약 ID가 존재하지 않아요.`                         |
+| 삭제하려는 예약이 존재하지 않을 경우         | 400 Bad Request | `존재하지 않는 예약이에요.`                           |
 
 ### 예약 조회 (GET /reservations)
 
@@ -42,3 +46,45 @@
 ### 예약 삭제 (DELETE /reservations/{id})
 
 예약 ID로 삭제
+
+---
+
+## 시간 API 사용법
+
+### 시간 추가 (POST /times)
+
+원하는 시간을 입력 시 추가 가능
+
+#### 요청 예시
+
+```json
+{
+  "time": "15:00"
+}
+```
+
+#### 응답 예시
+
+```json
+Location: /times/1
+
+{
+"id": 1,
+"time": "15:00"
+}
+```
+
+#### 예외 상황
+
+| 상황                   | 상태              | 예외 메세지                             |
+|----------------------|-----------------|------------------------------------|
+| 주어진 시간 형식에 맞지 않을 경우  | 400 Bad Request | `시간(시:분)형식에 맞게 입력해 주세요. ex) 15:30` |
+| 삭제하려는 시간이 존재하지 않을 경우 | 400 Bad Request | `존재하지 않는 시간이에요.`                   |
+
+### 시간 조회 (GET /times)
+
+모든 시간 목록 반환
+
+### 시간 삭제 (DELETE /times/{id})
+
+시간 ID로 삭제
