+++ src/main/java/roomescape/controller/page/ReservationPageController.java
@@ -1,0 +1,15 @@
+package roomescape.controller.page;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/reservation")
+public class ReservationPageController {
+
+    @GetMapping
+    public String reservation() {
+        return "reservation";
+    }
+}
