+++ src/main/java/roomescape/controller/page/TimePageController.java
@@ -1,15 +1,15 @@
-package roomescape.controller;
+package roomescape.controller.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("reservation")
-public class ReservationPageController {
+@RequestMapping("/time")
+public class TimePageController {
 
     @GetMapping
-    public String reservation() {
-        return "reservation";
+    public String time() {
+        return "time";
     }
 }
